requirements.txt
pandas>=1.3.0
openpyxl>=3.0.0
pathlib
argparse
# For bracket generation and PDF creation
reportlab>=3.6.0
pillow>=8.0.0
opencv-python>=4.5.0
flask>=2.0.0
numpy
Pillow
requests 